:construction: begin refactoring DCPR auth and actions
- Follow CKAN lead and provide initial dictization function
- Separate actions into create, delete, get modules
- Create dedicated actions for different user roles
- Create auth function skeletons
